requirements.txt
python-telegram-bot==20.7
telethon==1.28.5
aiosqlite==0.19.0